rename files to lower camel case
