+++ bin/cdk.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 const cdk = require("@aws-cdk/core");
-const { DynamoStack } = require("../lib/cdk-stack");
+const { DynamoStack } = require("../lib/cdkStack");
 
 const app = new cdk.App();
 new DynamoStack(app, "DynamoStack");
